backend/requirements.txt
asgiref==3.11.0
attrs==25.4.0
cachetools==6.2.2
certifi==2025.11.12
channels==4.2.0
channels-redis
charset-normalizer==3.4.4
Django==4.2.27
djangorestframework==3.16.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.12.1
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
googleapis-common-protos==1.72.0
httplib2==0.31.0
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mistralai==0.3.0
proto-plus==1.26.1
protobuf==6.33.1
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.5
python-decouple==3.8
PyYAML==6.0.3
referencing==0.36.2
redis==5.1.1
requests==2.32.5
rpds-py==0.27.1
rsa==4.9.1
sqlparse==0.5.4
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn[standard]==0.38.0
websockets==15.0.1
uvicorn==0.38.0
